Add ProcessTitle and User selector components
