+++ src/Components/Pages/Process/ProcessTitle.tsx
@@ -1,0 +1,8 @@
+export const ProcessTitle = () => {
+  return(
+    <>
+      <h1 className="heading">Terms of Service Processor</h1>
+      <p className="sub-heading">Understand what's important to you.</p>
+    </>
+  )
+}
